requirements.txt
flask==2.3.3
flask_session==0.5.0
langchain-openai==0.0.2
openai==1.3.0
python-dotenv==1.0.0
pytz==2023.3
langchain==0.0.267
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
Werkzeug==2.3.7